New function to calculate walking period using rolling average
The new method (in lines 51-56) creates a list of period values of length 3, and appends the newest value to the list as soon as it's created. This automatically gets rid of the oldest value. Currently only coded for right heel. Can be tested using the tester in the /tests folder.